requirements.txt
google-auth
google-cloud
google-cloud-monitoring
google-cloud-logging
google-cloud-pubsub